ai youtube tutor/requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
pytube
python-dotenv
youtube-transcript-api
sentence-transformers
faiss-cpu
google-generativeai